refactor(cli): extract shared validateAndParseHookConfig helper
The read/parse/assert-not-string/isHookCompatible/registry-lookup block was duplicated between `hook deploy` and `hook apply`; it now lives in a shared validateAndParseHookConfig helper so the two cannot drift. The deploy-specific unsupported-type check stays in deploy.ts, and the deploy parse-error message was unified to lowercase "Invalid hook config". Also added the cli package to the IGP fee integration changeset, covering the new hook deploy/read/apply commands.
