+++ .changeset/svm-igp-fee-integration.md
@@ -1,6 +1,7 @@
 ---
 '@hyperlane-xyz/sealevel-sdk': major
 '@hyperlane-xyz/provider-sdk': patch
+'@hyperlane-xyz/cli': minor
 ---
 
 SVM IGP fee config integration was added to the SVM SDK. SvmIgpHookReader was updated to surface the on-chain Igp.fee_config (signers + domainId + minIssuedAt) via the new feeConfig field on SvmDeployedIgpHook, and to expose the signer list through provider-sdk's IgpHookConfig.quoteSigners. SvmIgpHookWriter.create() and update() were updated to reconcile the multi-VM quoteSigners shape with on-chain state, mirroring EVM IGP semantics (undefined ⇒ leave on-chain state untouched, [] ⇒ keep fee_config Some without signers, [...] ⇒ initialize and/or Add/Remove diff). Clearing fee_config to None was intentionally left out of the declarative diff. The writer was wired to version-gate against the program's GetProgramVersion response (post-upgrade version when an upgrade fires in the same update) and to reject domain_id drift.
@@ -10,3 +11,5 @@ Breaking change: SvmIgpHookWriterConfig was extended to require domainId, and Sv
 Low-level codecs and instruction builders for the seven new IGP fee instructions were added: SetIgpQuoteConfig, SetIgpQuoteSigner, SetIgpMinIssuedAt, SubmitIgpQuote, CloseIgpTransientQuote, CloseIgpStandingQuote, and GetIgpQuoteAccountMetas. The IgpFeeConfig codec, IgpStandingQuote / IgpTransientQuote account decoders, the corresponding standing- and transient-quote PDA derivers, and WILDCARD_SENDER / WILDCARD_DOMAIN constants were promoted to public exports. SvmSignedQuote and GetIgpQuoteAccountMetasInput codecs were added with full encode/decode round-trip coverage.
 
 The provider-sdk IgpHookConfig was extended with optional contractVersion and quoteSigners fields, mirroring the EVM IgpSchema. Several previously private helpers were promoted to shared homes to support the new code without duplication: readAddress / readOptionAddress / ascii8 were moved into codecs/account-data.ts, and decodeBTreeSetH160 / decodeSetQuoteSignerOperation were colocated with the existing encoders in codecs/fee.ts. The svm-sdk unit-test runner glob was widened to src/**/*.unit-test.ts to pick up colocated codec / hook tests.
+
+The CLI gained `hyperlane hook deploy`, `hyperlane hook read`, and `hyperlane hook apply` commands, working across EVM and Alt-VM chains (including Sealevel IGP hooks with the fee config above). The shared read/parse/validate logic was extracted into a validateAndParseHookConfig helper so deploy and apply cannot drift, and the deploy path's config parse-error message was unified to lowercase "Invalid hook config".
