+++ typescript/cli/src/hook/apply.ts
@@ -2,11 +2,9 @@ import { createHookWriter } from '@hyperlane-xyz/deploy-sdk';
 import {
   EvmHookModule,
   type HookConfig,
-  HookConfigSchema,
   type TypedAnnotatedTransaction,
   altVmChainLookup,
   extractIsmAndHookFactoryAddresses,
-  isHookCompatible,
 } from '@hyperlane-xyz/sdk';
 import {
   type Address,
@@ -20,7 +18,8 @@ import { type WriteCommandContext } from '../context/types.js';
 import { validateHookConfigForAltVM } from '../deploy/configValidation.js';
 import { getSubmitterByStrategy } from '../deploy/warp.js';
 import { logCommandHeader, logGray, logGreen } from '../logger.js';
-import { readYamlOrJson } from '../utils/files.js';
+
+import { validateAndParseHookConfig } from './config.js';
 
 interface HookApplyParams {
   context: WriteCommandContext;
@@ -46,34 +45,12 @@ export async function runHookApply({
 
   const { multiProvider, registry } = context;
 
-  // Read and validate hook config
-  const rawConfig = await readYamlOrJson(configPath);
-  const parseResult = HookConfigSchema.safeParse(rawConfig);
-  if (!parseResult.success) {
-    const firstIssue = parseResult.error.issues[0];
-    throw new Error(
-      `Invalid hook config: ${firstIssue.path.join('.')} => ${firstIssue.message}`,
-    );
-  }
-  const hookConfig: HookConfig = parseResult.data;
-
-  assert(
-    typeof hookConfig !== 'string',
-    'Hook config must be an object, not an address string',
-  );
-
-  const { technicalStack } = multiProvider.getChainMetadata(chain);
-  assert(
-    isHookCompatible({
-      hookType: hookConfig.type,
-      chainTechnicalStack: technicalStack,
-    }),
-    `Hook type ${hookConfig.type} is not compatible with chain ${chain} (technical stack: ${technicalStack})`,
-  );
-
-  const chainAddresses = await registry.getChainAddresses(chain);
-  assert(chainAddresses, `No registry addresses found for chain ${chain}`);
-  assert(chainAddresses.mailbox, `No mailbox address found for chain ${chain}`);
+  const { hookConfig, chainAddresses } = await validateAndParseHookConfig({
+    configPath,
+    chain,
+    multiProvider,
+    registry,
+  });
 
   const protocol = multiProvider.getProtocol(chain);
   const { address: resultAddress, transactions } = isEVMLike(protocol)
